add trait for navigation badge limit/ count
