+++ app/Filament/Server/Resources/BackupResource.php
@@ -19,6 +19,7 @@ use App\Traits\Filament\CanCustomizePages;
 use App\Traits\Filament\CanCustomizeRelations;
 use App\Traits\Filament\CanModifyForm;
 use App\Traits\Filament\CanModifyTable;
+use App\Traits\Filament\HasLimitBadge;
 use Filament\Facades\Filament;
 use Filament\Forms\Components\Checkbox;
 use Filament\Forms\Components\Placeholder;
@@ -44,6 +45,7 @@ class BackupResource extends Resource
     use CanCustomizeRelations;
     use CanModifyForm;
     use CanModifyTable;
+    use HasLimitBadge;
 
     protected static ?string $model = Backup::class;
 
@@ -53,31 +55,20 @@ class BackupResource extends Resource
 
     protected static bool $canCreateAnother = false;
 
-    public const WARNING_THRESHOLD = 0.7;
-
-    public static function getNavigationBadge(): string
+    protected static function getBadgeCount(): int
     {
         /** @var Server $server */
         $server = Filament::getTenant();
 
-        $limit = $server->backup_limit;
-
-        return $server->backups->count() . ($limit === 0 ? '' : ' / ' . $limit);
+        return $server->backups->count();
     }
 
-    public static function getNavigationBadgeColor(): ?string
+    protected static function getBadgeLimit(): int
     {
         /** @var Server $server */
         $server = Filament::getTenant();
 
-        $limit = $server->backup_limit;
-        $count = $server->backups->count();
-
-        if ($limit === 0) {
-            return null;
-        }
-
-        return $count >= $limit ? 'danger' : ($count >= $limit * self::WARNING_THRESHOLD ? 'warning' : 'success');
+        return $server->backup_limit;
     }
 
     public static function defaultForm(Form $form): Form
